requirements.txt
fastai==2.6.3
torch==1.11.0
torchvision==0.12.0
rasterio==1.3b1
python-multipart==0.0.5
uvicorn==0.17.6
numpy==1.23.0
pandas==1.4.3
fastapi[all]==0.78.0
python-decouple
sqlalchemy==1.4.39
databases==0.6.0
fastapi-pagination[sqlalchemy]==0.9.3